static/python/requirements.txt
numpy==2.1.1
pandas==2.2.3
polars-lts-cpu==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
tzdata==2024.1
